# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/chat/session": {
      "post": {
        "operationId": "ChatController_createOrGetSession",
        "summary": "创建或获取聊天会话",
        "description": "每个用户和宠物只有一个活跃会话，如果存在则返回，不存在则创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreateSessionResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chat/session/current": {
      "get": {
        "operationId": "ChatController_getCurrentSession",
        "summary": "获取当前活跃会话",
        "description": "获取指定宠物的当前活跃会话",
        "parameters": [
          {
            "name": "petId",
            "required": true,
            "in": "query",
            "description": "宠物ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ChatSessionDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "会话不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chat/message/text": {
      "post": {
        "operationId": "ChatController_sendTextMessage",
        "summary": "发送文字消息",
        "description": "发送文字消息并获取AI回复",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTextMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SendMessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chat/message/image": {
      "post": {
        "operationId": "ChatController_sendImageMessage",
        "summary": "发送图片消息",
        "description": "发送图片消息（支持图片+文字）并获取AI回复",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendImageMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SendMessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chat/message/voice": {
      "post": {
        "operationId": "ChatController_sendVoiceMessage",
        "summary": "发送语音消息",
        "description": "发送语音消息并获取AI语音回复",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVoiceMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SendMessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chat/messages": {
      "get": {
        "operationId": "ChatController_getChatHistory",
        "summary": "获取聊天历史",
        "description": "获取指定宠物的聊天消息历史",
        "parameters": [
          {
            "name": "petId",
            "required": true,
            "in": "query",
            "description": "宠物ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "偏移量",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ChatHistoryResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chat/session/{id}/end": {
      "put": {
        "operationId": "ChatController_endSession",
        "summary": "结束聊天会话",
        "description": "结束当前活跃的聊天会话",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "会话ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/EndSessionResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "会话不存在或已结束",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chat/sessions": {
      "get": {
        "operationId": "ChatController_getSessions",
        "summary": "获取会话列表",
        "description": "获取用户的所有聊天会话列表",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "偏移量",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SessionListResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Claw Step API",
    "description": "Claw Step 应用后端 API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "ResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误信息",
            "example": "请求参数错误"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CreateSessionResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话ID",
            "example": "uuid-string"
          },
          "startTime": {
            "format": "date-time",
            "type": "string",
            "description": "开始时间"
          },
          "petId": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          }
        },
        "required": [
          "sessionId",
          "startTime",
          "petId"
        ]
      },
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          }
        },
        "required": [
          "petId"
        ]
      },
      "ChatSessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "会话ID",
            "example": "uuid-string"
          },
          "userId": {
            "type": "string",
            "description": "用户ID",
            "example": "uuid-string"
          },
          "petId": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          },
          "startTime": {
            "format": "date-time",
            "type": "string",
            "description": "开始时间"
          },
          "endTime": {
            "format": "date-time",
            "type": "string",
            "description": "结束时间"
          },
          "duration": {
            "type": "number",
            "description": "会话时长（秒）",
            "example": 3600
          },
          "messageCount": {
            "type": "number",
            "description": "消息总数",
            "example": 50
          },
          "userMessageCount": {
            "type": "number",
            "description": "用户消息数",
            "example": 25
          },
          "petMessageCount": {
            "type": "number",
            "description": "宠物消息数",
            "example": 25
          },
          "isActive": {
            "type": "boolean",
            "description": "是否活跃",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          }
        },
        "required": [
          "id",
          "userId",
          "petId",
          "startTime",
          "duration",
          "messageCount",
          "userMessageCount",
          "petMessageCount",
          "isActive",
          "createdAt"
        ]
      },
      "ChatMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "消息ID",
            "example": "uuid-string"
          },
          "sessionId": {
            "type": "string",
            "description": "会话ID"
          },
          "type": {
            "type": "string",
            "description": "消息类型",
            "enum": [
              "userText",
              "userImage",
              "userImageText",
              "userVoice",
              "petText"
            ],
            "example": "userText"
          },
          "content": {
            "type": "string",
            "description": "消息内容",
            "example": "你好，今天天气怎么样？"
          },
          "mediaUrls": {
            "description": "媒体URLs",
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mediaType": {
            "type": "string",
            "description": "媒体类型",
            "example": "none"
          },
          "duration": {
            "type": "number",
            "description": "语音时长（秒）",
            "example": 5
          },
          "isFromUser": {
            "type": "boolean",
            "description": "是否来自用户",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          }
        },
        "required": [
          "id",
          "type",
          "content",
          "mediaUrls",
          "mediaType",
          "isFromUser",
          "createdAt"
        ]
      },
      "SendMessageResponseDto": {
        "type": "object",
        "properties": {
          "userMessage": {
            "description": "用户消息",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatMessageDto"
              }
            ]
          },
          "petMessage": {
            "description": "宠物回复消息",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatMessageDto"
              }
            ]
          },
          "sessionId": {
            "type": "string",
            "description": "会话ID",
            "example": "uuid-string"
          },
          "petVoiceUrl": {
            "type": "string",
            "description": "宠物语音URL（如果是语音消息）"
          }
        },
        "required": [
          "userMessage",
          "petMessage"
        ]
      },
      "SendTextMessageDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话ID（可选，如果不提供会自动创建或获取）",
            "example": "uuid-string"
          },
          "petId": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          },
          "content": {
            "type": "string",
            "description": "消息内容",
            "example": "你好，今天天气怎么样？"
          }
        },
        "required": [
          "petId",
          "content"
        ]
      },
      "SendImageMessageDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话ID（可选）",
            "example": "uuid-string"
          },
          "petId": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          },
          "imageUrl": {
            "type": "string",
            "description": "图片URL",
            "example": "https://example.com/image.jpg"
          },
          "content": {
            "type": "string",
            "description": "文字内容（可选）",
            "example": "看看这张照片"
          }
        },
        "required": [
          "petId",
          "imageUrl"
        ]
      },
      "SendVoiceMessageDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话ID（可选）",
            "example": "uuid-string"
          },
          "petId": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          },
          "voiceUrl": {
            "type": "string",
            "description": "语音文件URL",
            "example": "https://example.com/voice.mp3"
          },
          "duration": {
            "type": "number",
            "description": "语音时长（秒）",
            "example": 5
          }
        },
        "required": [
          "petId",
          "voiceUrl",
          "duration"
        ]
      },
      "ChatHistoryResponseDto": {
        "type": "object",
        "properties": {
          "messages": {
            "description": "消息列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageDto"
            }
          },
          "total": {
            "type": "number",
            "description": "总数量",
            "example": 50
          },
          "sessionId": {
            "type": "string",
            "description": "会话ID"
          }
        },
        "required": [
          "messages",
          "total"
        ]
      },
      "EndSessionResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话ID",
            "example": "uuid-string"
          },
          "duration": {
            "type": "number",
            "description": "会话时长（秒）",
            "example": 3600
          },
          "messageCount": {
            "type": "number",
            "description": "消息总数",
            "example": 50
          }
        },
        "required": [
          "sessionId",
          "duration",
          "messageCount"
        ]
      },
      "SessionListResponseDto": {
        "type": "object",
        "properties": {
          "sessions": {
            "description": "会话列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatSessionDto"
            }
          },
          "total": {
            "type": "number",
            "description": "总数量",
            "example": 10
          }
        },
        "required": [
          "sessions",
          "total"
        ]
      },
...
    }
  }
}